requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
pyyaml>=6.0.1
requests>=2.31.0
sse-starlette>=1.8.2 
